requirements.txt
asgiref==3.5.1
Babel==2.10.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Django==3.2
django-appconf==1.0.5
django-autoslug==1.9.8
django-cors-headers==3.11.0
django-extra-views==0.13.0
django-filter==21.1
django-haystack==3.2.1
django-imagekit==4.1.0
django-oscar==3.1
django-oscar-api==3.1.1
django-phonenumber-field==5.2.0
django-tables2==2.3.4
django-templated-mail==1.1.1
django-treebeard==4.4
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
factory-boy==3.2.1
Faker==13.11.1
flake8==4.0.1
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
phonenumbers==8.12.48
pilkit==2.0
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
purl==1.6
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytils==0.4.1
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
